webapp: update delete_junk.py script

+++ flydra/sge_utils/delete_junk.py
@@ -4,17 +4,41 @@ import sys, os, shutil
 def delete_junk(couch_url, db_name, datasink_dir):
     couch_server = Server(couch_url)
     db = couch_server[db_name]
-    view_results = db.view('analysis/empty')
+    view_results = db.view('analysis/jobs')
+    DELETE = False
     for row in view_results:
         print row.id
-        print row.value
+        doc = db[row.id]
+        print doc
         print
 
         jobdir = os.path.join(datasink_dir, row.id)
-        if os.path.exists( jobdir ):
-            shutil.rmtree( jobdir )
+        if DELETE:
+            if os.path.exists( jobdir ):
+                shutil.rmtree( jobdir )
 
-        db.delete( row.value )
+            db.delete( doc )
+    startkey=["dataset:humdra_200809", "3d position"]
+    endkey=["dataset:humdra_200809", "3d position",{}]
+    view_results = db.view('analysis/datanodes-by-dataset-and-property',
+                           startkey=startkey,
+                           endkey=endkey,
+                           reduce=False,
+                           )
+    for row in view_results:
+        print row.id
+        if row.id is None:
+            continue
+        doc = db[row.id]
+        print doc
+        print
+
+        jobdir = os.path.join(datasink_dir, row.id)
+        if DELETE:
+            if os.path.exists( jobdir ):
+                shutil.rmtree( jobdir )
+
+            db.delete( doc )
 
 def main():
     couch_url = sys.argv[1]
